refactoring : line에 존재하는 소스 일부를 sections 로 이동
